+++ detectors/node/opentelemetry-resource-detector-aws/test/detectors/AwsEcsDetector.test.ts
@@ -199,10 +199,10 @@ describe('AwsEcsResourceDetector', () => {
       process.env.ECS_CONTAINER_METADATA_URI_V4 = ECS_CONTAINER_METADATA_URI_V4;
     });
 
-    describe('when succesfully retrieving the data', () => {
+    describe('when successfully retrieving the data', () => {
       function generateLaunchTypeTests(
         resourceAttributes: EcsResourceAttributes,
-        suffix: String = ''
+        suffix = ''
       ) {
         let nockScope: nock.Scope;