+++ plugins/node/opentelemetry-instrumentation-express/src/instrumentation.ts
@@ -39,7 +39,7 @@ import {
   isWrapped,
   safeExecuteInTheMiddle,
 } from '@opentelemetry/instrumentation';
-import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
+import { SEMATTRS_HTTP_ROUTE } from '@opentelemetry/semantic-conventions';
 import {
   ExpressLayer,
   ExpressRouter,
@@ -197,10 +197,12 @@ export class ExpressInstrumentation extends InstrumentationBase<
         storeLayerPath(req, layerPath);
         const route = (req[_LAYERS_STORE_PROPERTY] as string[])
           .filter(path => path !== '/' && path !== '/*')
-          .join('');
+          .join('')
+          // remove duplicate slashes to normalize route
+          .replace(/\/{2,}/g, '/');
 
         const attributes: Attributes = {
-          [SemanticAttributes.HTTP_ROUTE]: route.length > 0 ? route : '/',
+          [SEMATTRS_HTTP_ROUTE]: route.length > 0 ? route : '/',
         };
         const metadata = getLayerMetadata(layer, layerPath);
         const type = metadata.attributes[