+++ plugins/node/opentelemetry-instrumentation-express/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.36.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-express-v0.36.0...instrumentation-express-v0.36.1) (2024-03-11)
+
+
+### Bug Fixes
+
+* **instr-express:** normalize paths with double slashes ([#1995](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1995)) ([65a9553](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/65a9553e76a3e61da71c31758b6e5320f286374b))
+
+## [0.36.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-express-v0.35.0...instrumentation-express-v0.36.0) (2024-03-06)
+
+
+### Features
+
+* **deps:** update otel-js to 1.22.0/0.49.1 ([edc426b](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/edc426b348bc5f45ff6816bcd5ea7473251a05df))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.36.0 to ^0.37.0
+
 ## [0.35.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-express-v0.34.1...instrumentation-express-v0.35.0) (2024-01-29)